refactor: consolidate tour pages and remove duplicate code
