+++ src/components/pages/tours/GroupTourCard/index.tsx
@@ -4,9 +4,9 @@ import { useRef, useState, useEffect } from 'react'
 import Image from 'next/image'
 import { cn } from '@/lib/utils'
 import { RightOnlyFlipBook } from '@/components/ui/RightOnlyFlipBook'
-import type { TBaseComponent, TFlipBookPage } from '@/types'
+import type { TBaseComponent, TFlipBookPage } from '../../../../types'
 import styles from './style.module.css'
-import type { TTourDate } from './config'
+import type { TTourDate } from '../config'
 
 type TGroupTourCardProps = TBaseComponent & {
   title: string
