+++ src/components/pages/tours/TourBanner/index.tsx
